Removed unused ClientExec, moved ClientExecReturn into replication code
